Rename turn test for compliance
